refactor: split sio-registry.ts into focused modules
Split the 1,530-line sio-registry.ts into five focused modules under a
new sio-registry/ subdirectory:

- context.ts   — shared io ref, local socket Maps, room helpers, emit
                 helpers (emitToRunner/Relay), safeJsonParse,
                 nextEphemeralExpiry, modelFromHeartbeat
- hub.ts        — broadcastToHub, addHubClient, removeHubClient
- sessions.ts   — session + viewer lifecycle, heartbeat, sweep (~650 lines)
- runners.ts    — runner registration, skills/agents/plugins, session links
- terminals.ts  — terminal register/viewer/buffer/GC

sio-registry.ts is now a one-liner re-export barrel; all 7 existing
callers continue to import from the same path with zero changes.

No logic changes — pure reorganization.
